# airlock-swagger-client/src/main/resources/airlock-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": ""
  },
  "basePath": "/airlock/api",
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Retrieves all products.",
        "description": "",
        "operationId": "getAllProducts",
        "produces": [
          "application/json"
        ],
        "parameters": [

        ],
        "responses": {
          "200": {
            "description": "Retrieve all products completed successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "Product name."
        },
        "productId": {
          "type": "string",
          "description": "Product id."
        },
        "seasonId": {
          "type": "string",
          "description": "The season of the application associated with this product as it was provided during initialization."
        },
        "appVersion": {
          "type": "string",
          "description": "The version of the application associated with this product as it was provided during initialization."
        },
        "instanceId": {
          "type": "string",
          "description": "The instanceId of this product as it was provided during initialization."
        }
      },
      "description": "Represents an Airlock product registered on the SDK."
    },
...
  }
}